Big warnings cleanup
Removed the "first level" of warnings. During the phase, some other bugs
have been fixed. Game is still not working, a bigger cleanup is needed,
mostly related to those *damn* casts.

Jpeg lib sources have been removed.

+++ d3xp/gamesys/Class.h
@@ -59,7 +59,7 @@ class idEventArg
 {
 public:
     int			type;
-    int			value;
+    ptrdiff_t	value;
 
     idEventArg()
     {
